+++ jobbie-admin/electron/preload.cjs
@@ -1,5 +1,7 @@
-const { contextBridge } = require('electron')
+const { contextBridge, ipcRenderer } = require('electron')
 
 contextBridge.exposeInMainWorld('jobbieAdmin', {
   platform: process.platform,
+  getApiBootstrapStatus: () => ipcRenderer.invoke('admin:getApiBootstrapStatus'),
+  openUserDataFolder: () => ipcRenderer.invoke('admin:openUserDataFolder'),
 })